% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\docType{class}
\name{Animal}
\alias{Animal}
\title{Animal}
\format{
An \code{R6Class} generator object
}
\description{
Animal Class

Animal Class
}
\details{
Create a new Animal
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{className}}{character}

\item{\code{color}}{character [optional]}

\item{\code{_field_list}}{a list of fields list(character)}

\item{\code{additional_properties}}{additional properties list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Animal-new}{\code{Animal$new()}}
\item \href{#method-Animal-toJSON}{\code{Animal$toJSON()}}
\item \href{#method-Animal-fromJSON}{\code{Animal$fromJSON()}}
\item \href{#method-Animal-toJSONString}{\code{Animal$toJSONString()}}
\item \href{#method-Animal-fromJSONString}{\code{Animal$fromJSONString()}}
\item \href{#method-Animal-validateJSON}{\code{Animal$validateJSON()}}
\item \href{#method-Animal-toString}{\code{Animal$toString()}}
\item \href{#method-Animal-isValid}{\code{Animal$isValid()}}
\item \href{#method-Animal-getInvalidFields}{\code{Animal$getInvalidFields()}}
\item \href{#method-Animal-print}{\code{Animal$print()}}
\item \href{#method-Animal-clone}{\code{Animal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Animal class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$new(className, color = "red", additional_properties = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{className}

\item{\code{color}}{color. Default to "red".}

\item{\code{additional_properties}}{additonal properties (optional)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Animal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Animal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Animal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Animal in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Animal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Animal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Animal and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Animal
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-print"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Animal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Animal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
