% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_of_primitive_type_test.R
\docType{class}
\name{AnyOfPrimitiveTypeTest}
\alias{AnyOfPrimitiveTypeTest}
\title{AnyOfPrimitiveTypeTest}
\format{
An \code{R6Class} generator object
}
\description{
AnyOfPrimitiveTypeTest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new AnyOfPrimitiveTypeTest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnyOfPrimitiveTypeTest-new}{\code{AnyOfPrimitiveTypeTest$new()}}
\item \href{#method-AnyOfPrimitiveTypeTest-fromJSONString}{\code{AnyOfPrimitiveTypeTest$fromJSONString()}}
\item \href{#method-AnyOfPrimitiveTypeTest-fromJSON}{\code{AnyOfPrimitiveTypeTest$fromJSON()}}
\item \href{#method-AnyOfPrimitiveTypeTest-toJSONString}{\code{AnyOfPrimitiveTypeTest$toJSONString()}}
\item \href{#method-AnyOfPrimitiveTypeTest-toJSON}{\code{AnyOfPrimitiveTypeTest$toJSON()}}
\item \href{#method-AnyOfPrimitiveTypeTest-validateJSON}{\code{AnyOfPrimitiveTypeTest$validateJSON()}}
\item \href{#method-AnyOfPrimitiveTypeTest-toString}{\code{AnyOfPrimitiveTypeTest$toString()}}
\item \href{#method-AnyOfPrimitiveTypeTest-print}{\code{AnyOfPrimitiveTypeTest$print()}}
\item \href{#method-AnyOfPrimitiveTypeTest-clone}{\code{AnyOfPrimitiveTypeTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new AnyOfPrimitiveTypeTest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "character", "integer"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of AnyOfPrimitiveTypeTest.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AnyOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of AnyOfPrimitiveTypeTest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AnyOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize AnyOfPrimitiveTypeTest to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the AnyOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize AnyOfPrimitiveTypeTest to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the AnyOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to AnyOfPrimitiveTypeTest and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-print"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnyOfPrimitiveTypeTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnyOfPrimitiveTypeTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfPrimitiveTypeTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
