% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basque_pig.R
\docType{class}
\name{BasquePig}
\alias{BasquePig}
\title{BasquePig}
\format{
An \code{R6Class} generator object
}
\description{
BasquePig Class

BasquePig Class
}
\details{
Create a new BasquePig
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{className}}{character}

\item{\code{color}}{character}

\item{\code{_field_list}}{a list of fields list(character)}

\item{\code{additional_properties}}{additional properties list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasquePig-new}{\code{BasquePig$new()}}
\item \href{#method-BasquePig-toJSON}{\code{BasquePig$toJSON()}}
\item \href{#method-BasquePig-fromJSON}{\code{BasquePig$fromJSON()}}
\item \href{#method-BasquePig-toJSONString}{\code{BasquePig$toJSONString()}}
\item \href{#method-BasquePig-fromJSONString}{\code{BasquePig$fromJSONString()}}
\item \href{#method-BasquePig-validateJSON}{\code{BasquePig$validateJSON()}}
\item \href{#method-BasquePig-toString}{\code{BasquePig$toString()}}
\item \href{#method-BasquePig-isValid}{\code{BasquePig$isValid()}}
\item \href{#method-BasquePig-getInvalidFields}{\code{BasquePig$getInvalidFields()}}
\item \href{#method-BasquePig-print}{\code{BasquePig$print()}}
\item \href{#method-BasquePig-clone}{\code{BasquePig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new BasquePig class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$new(className, color, additional_properties = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{className}

\item{\code{color}}{color}

\item{\code{additional_properties}}{additonal properties (optional)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BasquePig in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of BasquePig
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BasquePig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BasquePig in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of BasquePig
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of BasquePig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to BasquePig and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-toString"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of BasquePig
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-print"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasquePig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasquePig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasquePig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
