% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammal.R
\docType{class}
\name{Mammal}
\alias{Mammal}
\title{Mammal}
\format{
An \code{R6Class} generator object
}
\description{
Mammal Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new Mammal.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mammal-new}{\code{Mammal$new()}}
\item \href{#method-Mammal-fromJSONString}{\code{Mammal$fromJSONString()}}
\item \href{#method-Mammal-fromJSON}{\code{Mammal$fromJSON()}}
\item \href{#method-Mammal-toJSONString}{\code{Mammal$toJSONString()}}
\item \href{#method-Mammal-toJSON}{\code{Mammal$toJSON()}}
\item \href{#method-Mammal-validateJSON}{\code{Mammal$validateJSON()}}
\item \href{#method-Mammal-toString}{\code{Mammal$toString()}}
\item \href{#method-Mammal-print}{\code{Mammal$print()}}
\item \href{#method-Mammal-clone}{\code{Mammal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Mammal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "Whale", "Zebra"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Mammal.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of Mammal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Mammal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of Mammal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize Mammal to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the Mammal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize Mammal to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the Mammal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to Mammal and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-print"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mammal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Mammal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mammal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
