% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api_response.R
\docType{class}
\name{ModelApiResponse}
\alias{ModelApiResponse}
\title{ModelApiResponse}
\format{
An \code{R6Class} generator object
}
\description{
Describes the result of uploading an image resource

ModelApiResponse Class
}
\details{
Create a new ModelApiResponse
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{integer [optional]}

\item{\code{type}}{character [optional]}

\item{\code{message}}{character [optional]}

\item{\code{_field_list}}{a list of fields list(character)}

\item{\code{additional_properties}}{additional properties list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelApiResponse-new}{\code{ModelApiResponse$new()}}
\item \href{#method-ModelApiResponse-toJSON}{\code{ModelApiResponse$toJSON()}}
\item \href{#method-ModelApiResponse-fromJSON}{\code{ModelApiResponse$fromJSON()}}
\item \href{#method-ModelApiResponse-toJSONString}{\code{ModelApiResponse$toJSONString()}}
\item \href{#method-ModelApiResponse-fromJSONString}{\code{ModelApiResponse$fromJSONString()}}
\item \href{#method-ModelApiResponse-validateJSON}{\code{ModelApiResponse$validateJSON()}}
\item \href{#method-ModelApiResponse-toString}{\code{ModelApiResponse$toString()}}
\item \href{#method-ModelApiResponse-isValid}{\code{ModelApiResponse$isValid()}}
\item \href{#method-ModelApiResponse-getInvalidFields}{\code{ModelApiResponse$getInvalidFields()}}
\item \href{#method-ModelApiResponse-print}{\code{ModelApiResponse$print()}}
\item \href{#method-ModelApiResponse-clone}{\code{ModelApiResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new ModelApiResponse class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$new(
  code = NULL,
  type = NULL,
  message = NULL,
  additional_properties = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{code}

\item{\code{type}}{type}

\item{\code{message}}{message}

\item{\code{additional_properties}}{additonal properties (optional)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ModelApiResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of ModelApiResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ModelApiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
ModelApiResponse in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of ModelApiResponse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of ModelApiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to ModelApiResponse and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-toString"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of ModelApiResponse
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelApiResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelApiResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelApiResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
