% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_of_primitive_type_test.R
\docType{class}
\name{OneOfPrimitiveTypeTest}
\alias{OneOfPrimitiveTypeTest}
\title{OneOfPrimitiveTypeTest}
\format{
An \code{R6Class} generator object
}
\description{
OneOfPrimitiveTypeTest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actual_instance}}{the object stored in this instance.}

\item{\code{actual_type}}{the type of the object stored in this instance.}

\item{\code{one_of}}{a list of types defined in the oneOf schema.
Initialize a new OneOfPrimitiveTypeTest.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OneOfPrimitiveTypeTest-new}{\code{OneOfPrimitiveTypeTest$new()}}
\item \href{#method-OneOfPrimitiveTypeTest-fromJSONString}{\code{OneOfPrimitiveTypeTest$fromJSONString()}}
\item \href{#method-OneOfPrimitiveTypeTest-fromJSON}{\code{OneOfPrimitiveTypeTest$fromJSON()}}
\item \href{#method-OneOfPrimitiveTypeTest-toJSONString}{\code{OneOfPrimitiveTypeTest$toJSONString()}}
\item \href{#method-OneOfPrimitiveTypeTest-toJSON}{\code{OneOfPrimitiveTypeTest$toJSON()}}
\item \href{#method-OneOfPrimitiveTypeTest-validateJSON}{\code{OneOfPrimitiveTypeTest$validateJSON()}}
\item \href{#method-OneOfPrimitiveTypeTest-toString}{\code{OneOfPrimitiveTypeTest$toString()}}
\item \href{#method-OneOfPrimitiveTypeTest-print}{\code{OneOfPrimitiveTypeTest$print()}}
\item \href{#method-OneOfPrimitiveTypeTest-clone}{\code{OneOfPrimitiveTypeTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new OneOfPrimitiveTypeTest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$new(instance = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{an instance of the object defined in the oneOf schemas: "character", "integer"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of OneOfPrimitiveTypeTest.
An alias to the method `fromJSON` .
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$fromJSONString(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OneOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of OneOfPrimitiveTypeTest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$fromJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of OneOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
Serialize OneOfPrimitiveTypeTest to JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON string representation of the OneOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
Serialize OneOfPrimitiveTypeTest to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON representation of the OneOfPrimitiveTypeTest.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate the input JSON with respect to OneOfPrimitiveTypeTest and
throw exception if invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input JSON.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-toString"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-toString}{}}}
\subsection{Method \code{toString()}}{
Returns the string representation of the instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The string representation of the instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-print"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OneOfPrimitiveTypeTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OneOfPrimitiveTypeTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneOfPrimitiveTypeTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
