% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\docType{class}
\name{Order}
\alias{Order}
\title{Order}
\format{
An \code{R6Class} generator object
}
\description{
An order for a pets from the pet store

Order Class
}
\details{
Create a new Order
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{petId}}{integer [optional]}

\item{\code{quantity}}{integer [optional]}

\item{\code{shipDate}}{character [optional]}

\item{\code{status}}{Order Status character [optional]}

\item{\code{complete}}{character [optional]}

\item{\code{_field_list}}{a list of fields list(character)}

\item{\code{additional_properties}}{additional properties list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Order-new}{\code{Order$new()}}
\item \href{#method-Order-toJSON}{\code{Order$toJSON()}}
\item \href{#method-Order-fromJSON}{\code{Order$fromJSON()}}
\item \href{#method-Order-toJSONString}{\code{Order$toJSONString()}}
\item \href{#method-Order-fromJSONString}{\code{Order$fromJSONString()}}
\item \href{#method-Order-validateJSON}{\code{Order$validateJSON()}}
\item \href{#method-Order-toString}{\code{Order$toString()}}
\item \href{#method-Order-isValid}{\code{Order$isValid()}}
\item \href{#method-Order-getInvalidFields}{\code{Order$getInvalidFields()}}
\item \href{#method-Order-print}{\code{Order$print()}}
\item \href{#method-Order-clone}{\code{Order$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-new"></a>}}
\if{latex}{\out{\hypertarget{method-Order-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Order class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$new(
  id = NULL,
  petId = NULL,
  quantity = NULL,
  shipDate = NULL,
  status = NULL,
  complete = FALSE,
  additional_properties = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{petId}}{petId}

\item{\code{quantity}}{quantity}

\item{\code{shipDate}}{shipDate}

\item{\code{status}}{Order Status}

\item{\code{complete}}{complete. Default to FALSE.}

\item{\code{additional_properties}}{additonal properties (optional)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Order-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Order in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Order-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Order
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Order-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Order in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Order-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Order
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Order-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Order and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Order-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Order
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Order-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Order-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-print"></a>}}
\if{latex}{\out{\hypertarget{method-Order-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Order-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Order-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Order$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
