% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petstore_api.R
\docType{class}
\name{petstore_api}
\alias{petstore_api}
\title{ApiClient}
\format{
An \code{R6Class} generator object
}
\description{
ApiClient Class
}
\details{
OpenAPI Petstore

This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.

The version of the OpenAPI document: 1.0.0
Generated by: https://openapi-generator.tech

petstore API Class

A single point of access to the petstore API.

NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
Ref: https://openapi-generator.tech
Do not edit the class manually.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_client}}{API client}

\item{\code{fake_api}}{an instance of FakeApi}

\item{\code{pet_api}}{an instance of PetApi}

\item{\code{store_api}}{an instance of StoreApi}

\item{\code{user_api}}{an instance of UserApi}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-petstore_api-new}{\code{petstore_api$new()}}
\item \href{#method-petstore_api-clone}{\code{petstore_api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-petstore_api-new"></a>}}
\if{latex}{\out{\hypertarget{method-petstore_api-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new petstore API Class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{petstore_api$new(api_client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_client}}{An instance of API client (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-petstore_api-clone"></a>}}
\if{latex}{\out{\hypertarget{method-petstore_api-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{petstore_api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
