% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\docType{class}
\name{Tag}
\alias{Tag}
\title{Tag}
\format{
An \code{R6Class} generator object
}
\description{
A tag for a pet

Tag Class
}
\details{
Create a new Tag
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{integer [optional]}

\item{\code{name}}{character [optional]}

\item{\code{_field_list}}{a list of fields list(character)}

\item{\code{additional_properties}}{additional properties list(character) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tag-new}{\code{Tag$new()}}
\item \href{#method-Tag-toJSON}{\code{Tag$toJSON()}}
\item \href{#method-Tag-fromJSON}{\code{Tag$fromJSON()}}
\item \href{#method-Tag-toJSONString}{\code{Tag$toJSONString()}}
\item \href{#method-Tag-fromJSONString}{\code{Tag$fromJSONString()}}
\item \href{#method-Tag-validateJSON}{\code{Tag$validateJSON()}}
\item \href{#method-Tag-toString}{\code{Tag$toString()}}
\item \href{#method-Tag-isValid}{\code{Tag$isValid()}}
\item \href{#method-Tag-getInvalidFields}{\code{Tag$getInvalidFields()}}
\item \href{#method-Tag-print}{\code{Tag$print()}}
\item \href{#method-Tag-clone}{\code{Tag$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Tag class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$new(id = NULL, name = NULL, additional_properties = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{name}}{name}

\item{\code{additional_properties}}{additonal properties (optional)}

\item{\code{...}}{Other optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$toJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Tag in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
Deserialize JSON string into an instance of Tag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$fromJSON(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Tag
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
To JSON String
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$toJSONString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Tag in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
Deserialize JSON string into an instance of Tag
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$fromJSONString(input_json)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_json}}{the JSON input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the instance of Tag
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-validateJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-validateJSON}{}}}
\subsection{Method \code{validateJSON()}}{
Validate JSON input with respect to Tag and throw an exception if invalid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$validateJSON(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the JSON input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-toString"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-toString}{}}}
\subsection{Method \code{toString()}}{
To string (JSON format)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$toString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String representation of Tag
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-isValid}{}}}
\subsection{Method \code{isValid()}}{
Return true if the values in all fields are valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$isValid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
true if the values in all fields are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-getInvalidFields"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-getInvalidFields}{}}}
\subsection{Method \code{getInvalidFields()}}{
Return a list of invalid fields (if any).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$getInvalidFields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of invalid fields (if any).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-print"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-print}{}}}
\subsection{Method \code{print()}}{
Print the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tag-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tag-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
